py/sage/requirements.txt
boto3==1.28.12
botocore==1.31.12
deepface==0.0.79
ffmpeg==1.4
ffmpeg-python
numpy==1.25.1
pandas==2.0.3
python-dotenv==1.0.0
tensorflow==2.13.0
torch==2.0.1
